refactor firebase client initialization
