+++ src/pages/meals.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { auth, db } from "@/lib/firebase";
+import { auth, db } from "@/lib/firebaseClient";
 import { onAuthStateChanged, signOut } from "firebase/auth";
 import { collection, addDoc, query, orderBy, onSnapshot, Timestamp } from "firebase/firestore";
